sat: Fix TDR max retries
We have to max cap a drift to 90 minutes transmission period.  Limiting
the number of retries to 255 is not enough to reach 90 minutes.

Just change the type of retries to uint16_t and let it holds enough
retries to reach the 90 min cap.

+++ include/hubble/port/sat_radio.h
@@ -93,7 +93,7 @@ int hubble_sat_port_init(void);
  * @return 0 on successful transmission, negative error code on failure.
  */
 int hubble_sat_port_packet_send(const struct hubble_sat_packet *packet,
-				uint8_t retries, uint8_t interval_s);
+				uint16_t retries, uint8_t interval_s);
 
 /**
  * @brief Transmit a DTM test packet over the satellite radio.
